Refactor API import in ProductContext and create api utility for axios instance
