+++ frontend/src/context/ProductContext.jsx
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, useState, useEffect } from "react";
-import api from "../utils/api";
+import api from "../lib/api";
 
 const ProductContext = createContext();
